Added a new approach (now the default)
The new approach tries to get everything below a series of lower and lower target rates. This can get rid of expensive borrowing fast, and capture any short lived cheap funding offers as it works the prices down

Also, rearranged a lot of code to support multiple stategies for replacing funding.

Also removed the config dependency and replaced it with a simple local version.