Now track position and wallets to calculate expected borrowing
Also tweaked the code to return existing unwanted borrowing, as occasionally 500 error were being reported. Not tracked down a reason, but now logs more useful data, so it might be easier to see a pattern.

Generally improved some of the logging output too, so more relevant data is presented

Since it is now able to display your wallet balance and all open positions, there is a new options to disable the display of that info.